updated FP examples

+++ code_samples/lec5/fp_precision.cpp
@@ -20,7 +20,7 @@ const unsigned long PRINT_INTERVAL = 1e+06;   // one million
 
 int main(int argc, char *argv[]) {
 
-  unsigned long i = 0; 
+  float i = 0; 
   int j = 0;
   while (i  <= LOOP_BOUND) {
     // print counter value every millionth iteration
